backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.3.0
openai==1.86.0
python-dotenv==1.1.0
slowapi==0.1.9
redis==5.2.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.28.1

